requirements.txt
fastapi
uvicorn
playwright
playwright-stealth
pydantic
python-dotenv
